third_party/WebKit/LayoutTests/external/wpt/bluetooth/requestDevice/not-processing-user-gesture.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => setUpHealthThermometerAndHeartRateDevices()
    .then(() => assert_promise_rejects_with_message(
        navigator.bluetooth.requestDevice({filters:[{services:['heart_rate']}]}),
        new DOMException(
          'Must be handling a user gesture to show a permission request.',
          'SecurityError'),
        'User gesture is required')),
    'Requires a user gesture.');
</script>
